requirements.txt
# === Runtime Dependencies ===
requests>=2.32.3
cryptography>=43.0.0
pyyaml>=6.0.0
pytz>=2024.2
python-box>=7.3.0
python-dateutil>=2.9.0
flatdict>=4.0.0
pyjwt>=2.8.0
aenum>=3.1.11
pydash>=8.0.3
arrow>=1.3.0

# === Optional Runtime Dependencies (if used elsewhere dynamically) ===
yarl>=1.9.0
xmltodict>=0.14.2
pycryptodomex>=3.20.0
jwcrypto>=1.4.2

# === Linting/Formatting ===
black>=24.3.0
flake8>=7.0.0
isort>=5.13.0

# === Testing Dependencies ===
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
pytest-recording>=0.13.2
pyfakefs>=5.7.0
responses>=0.25.0

# === Documentation ===
sphinx>=7.2.0
sphinx-autobuild>=2024.1.0
sphinx_rtd_theme>=1.3.0

# === Packaging ===
wheel>=0.42.0
